refactor: 플랫폼 UI 매핑을 provider.ts로 이동
